requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.2
bleach==3.1.4
boto3==1.13.1
botocore==1.16.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.2
dicttoxml==1.7.4
dj-database-url==0.5.0
Django==3.0.5
django-crontab==0.7.1
django-filter==2.2.0
django-ipware==2.1.0
django-qr-code==1.2.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
geocoder==1.38.1
geoip2==3.0.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
ip2geotools==0.1.5
IP2Location==8.4.1
isort==4.3.21
jeepney==0.4.3
jmespath==0.9.5
keyring==21.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
maxminddb==1.5.2
mccabe==0.6.1
packaging==20.3
Pillow==7.1.2
pip-review==1.1.0
pkginfo==1.5.0.1
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.3
qrcode==6.1
ratelim==0.1.6
readme-renderer==26.0
requests==2.23.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
SecretStorage==3.1.2
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
tqdm==4.45.0
twine==3.1.1
typed-ast==1.4.1
typing==3.7.4.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
